Move ajv stuff to utils
