+++ front/app/components/Form/utils.ts
@@ -1,4 +1,4 @@
-import { Layout, JsonSchema7 } from '@jsonforms/core';
+import { Layout, JsonSchema7, createAjv } from '@jsonforms/core';
 import Ajv from 'ajv';
 import { isEmpty, forOwn } from 'lodash-es';
 
@@ -123,3 +123,10 @@ export function getOtherControlKey(scope: string = ''): string | undefined {
 
   return undefined;
 }
+
+export const customAjv = createAjv({
+  useDefaults: 'empty',
+  removeAdditional: true,
+});
+// The image key word is used for the image choice option
+customAjv.addKeyword('image');
