+++ front/app/containers/Admin/projects/project/inputImporter/ReviewSection/IdeaEditor/index.tsx
@@ -25,8 +25,8 @@ import useLocalize from 'hooks/useLocalize';
 
 import { getFormValues as getIdeaFormValues } from 'containers/IdeasEditPage/utils';
 
-import { customAjv } from 'components/Form';
 import { FormData } from 'components/Form/typings';
+import { customAjv } from 'components/Form/utils';
 import { isValidData } from 'components/Form/utils';
 
 import { FormattedMessage } from 'utils/cl-intl';
